[Refactor] #615 - LegacyAuth, Auth 모듈 분리
