+++ SOPT-iOS/Plugins/DependencyPlugin/ProjectDescriptionHelpers/Dependency+Project.swift
@@ -22,6 +22,7 @@ public extension Dep {
         public struct Home {}
         public struct Soptlog {}
         public struct TabBar {}
+        public struct LegacyAuth {}
     }
     
     struct Modules {}
@@ -45,8 +46,6 @@ public extension Dep.Modules {
     static let networks = Dep.project(target: "Networks", path: .relativeToModules("Networks"))
     
     static let thirdPartyLibs = Dep.project(target: "ThirdPartyLibs", path: .relativeToModules("ThirdPartyLibs"))
-    
-    static let testCore = Dep.project(target: "TestCore", path: .relativeToModules("TestCore"))
 }
 
 // MARK: - Features
@@ -61,6 +60,13 @@ public extension Dep.Features {
 
 public extension Dep.Features.Spalsh {
     static let group = "Splash"
+    
+    static let Feature = Dep.Features.project(name: "Feature", group: group)
+    static let Interface = Dep.project(target: "\(group)FeatureInterface", path: .relativeToFeature("\(group)Feature"))
+}
+
+public extension Dep.Features.LegacyAuth {
+    static let group = "LegacyAuth"
     
     static let Feature = Dep.Features.project(name: "Feature", group: group)
     static let Interface = Dep.project(target: "\(group)FeatureInterface", path: .relativeToFeature("\(group)Feature"))
