+++ SOPT-iOS/Projects/Features/AuthFeature/Interface/Sources/SearchSocialAccountPresentable.swift
