+++ SOPT-iOS/Projects/Features/AuthFeature/Sources/Coordinator/AuthBuilder.swift
@@ -1,31 +1,80 @@
 //
-//  AuthBuilder.swift
-//  AuthFeatureTests
+//  AuthBuilder_Refactor.swift
+//  AuthFeature
 //
-//  Created by Junho Lee on 2023/06/19.
-//  Copyright © 2023 SOPT-iOS. All rights reserved.
+//  Created by 장석우 on 3/7/25.
+//  Copyright © 2025 SOPT-iOS. All rights reserved.
 //
 
 import Core
 import Domain
-import AuthFeatureInterface
 
-public class AuthBuilder: AuthFeatureViewBuildable {
+@_exported import AuthFeatureInterface
+import BaseFeatureDependency
+
+public final class AuthBuilder: AuthFeatureViewBuildable {
+    
     @Injected public var repository: SignInRepositoryInterface
     @Injected public var oauthRepository: CoreOAuthRepositoryInterface
     @Injected public var coreRepository: CoreAuthRepositoryInterface
+    @Injected public var phoneRepository: PhoneVerifyRepositoryInterface
     
     public init() { }
     
     public func makeSignIn() -> SignInPresentable {
-        let useCase = DefaultSignInUseCase(
-            repository: repository,
-            oauthRepository: oauthRepository,
-            coreRepository: coreRepository
-        )
-        let vm = SignInViewModel(useCase: useCase)
-        let vc = SignInVC()
+        let useCase = DefaultSignInUseCase(repository: repository,
+                                           oauthRepository: oauthRepository,
+                                           coreRepository: coreRepository)
+        let vm = SignInViewModel_Refactor(useCase: useCase)
+        let vc = SignInVC_Refactor()
         vc.viewModel = vm
         return (vc, vm)
     }
+    
+    public func makeLoginHelpBottomSheet() -> LoginHelpBottomSheetPresentable {
+        return LoginHelpBottomSheetVC()
+    }
+    
+    public func makeUserNotFound() -> UserNotFoundPresentable {
+        return UserNotFoundVC()
+    }
+    
+    public func makeSignUp() -> SignUpPresentable {
+        let useCase = DefaultSignUpUseCase(repository: coreRepository, oAuthRepository: oauthRepository)
+        let phoneUseCase = DefaultPhoneVerifyUseCase(repository: phoneRepository)
+        
+        let vm = SignUpViewModel(useCase: useCase)
+        let phoneVM = PhoneVerifyViewModel(useCase: phoneUseCase, phoneVerifyType: .register)
+        let vc = SignUpVC(viewModel: vm, phoneVerifyViewModel: phoneVM)
+        
+        return (vc, vm)
+    }
+    
+    public func makeChangeSocialAccount() -> ChangeSocialAccountPresentable {
+        let useCase = DefaultChangeSocialAccountUseCase(
+            repository: coreRepository,
+            oAuthRepository: oauthRepository
+        )
+        
+        let phoneUseCase = DefaultPhoneVerifyUseCase(repository: phoneRepository)
+        
+        let vm = ChangeSocialAccountViewModel(useCase: useCase)
+        let phoneVM = PhoneVerifyViewModel(useCase: phoneUseCase, phoneVerifyType: .changeSocialAccount)
+        let vc = ChangeSocialAccountVC(viewModel: vm, phoneVerifyViewModel: phoneVM)
+        return (vc, vm)
+    }
+    
+    public func makeSearchSocialAccount() -> SearchSocialAccountPresentable {
+        let useCase = DefaultSearchSocialAccountUseCase(
+            repository: coreRepository
+        )
+        
+        let phoneUseCase = DefaultPhoneVerifyUseCase(repository: phoneRepository)
+        
+        let vm = SearchSocialAccountViewModel(useCase: useCase)
+        let phoneVM = PhoneVerifyViewModel(useCase: phoneUseCase, phoneVerifyType: .searchSocialAccount)
+        let vc = SearchSocialAccountVC(viewModel: vm, phoneVerifyViewModel: phoneVM)
+        return (vc, vm)
+    }
+
 }
