+++ SOPT-iOS/Projects/Features/AuthFeature/Sources/SignInScene/SignInViewModel_Refactor.swift
@@ -38,7 +38,8 @@ public class SignInViewModel_Refactor: SignInViewModelType_Refactor {
     
     // MARK: - SignInCoordinating
     
-    public var onSignInSuccess: ((SiginInHandleableType) -> Void)?
+    // TODO: 신규 자동 로그인
+    public var onSignInSuccess: (() -> Void)?
     public var onLoginHelpButtonTapped: (() -> Void)?
     public var onVisitorButtonTapped: (() -> Void)?
     public var onSocialLoginFail: (() -> Void)?
@@ -54,7 +55,6 @@ public class SignInViewModel_Refactor: SignInViewModelType_Refactor {
 extension SignInViewModel_Refactor {
     public func transform(from input: Input, cancelBag: CancelBag) -> Output {
         let output = Output()
-        self.bindOutput(output: output, cancelBag: cancelBag)
         
         input.viewDidLoad
             .sink { _ in
@@ -77,7 +77,7 @@ extension SignInViewModel_Refactor {
         }
         .withUnretained(self)
         .sink { owner, _ in
-            owner.onSignInSuccess?(.loginSuccess)
+            owner.onSignInSuccess?()
         }.store(in: self.cancelBag)
           
         
@@ -100,16 +100,5 @@ extension SignInViewModel_Refactor {
             }.store(in: self.cancelBag)
         
         return output
-    }
-  
-    private func bindOutput(output: Output, cancelBag: CancelBag) {
-        useCase.signInSuccess
-            .removeDuplicates()
-            .withUnretained(self)
-            .sink { event in
-                print("SignInViewModel: \(event)")
-            } receiveValue: { (owner, isSignInSuccess) in
-                owner.onSignInSuccess?(isSignInSuccess)
-            }.store(in: self.cancelBag)
     }
 }
