+++ SOPT-iOS/Projects/Features/LegacyAuthFeature/Demo/Resources/Assets.xcassets/Contents.json
@@ -1,0 +1,6 @@
+{
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
