+++ SOPT-iOS/Projects/Features/LegacyAuthFeature/Tests/Resources/Empty.swift
@@ -1,0 +1,8 @@
+//
+//  Empty.swift
+//  Templates
+//
+//  Created by Melt on 06/06/25
+//
+
+import Foundation
